Adding the WatchedList pattern
Add the entity `watched-list`, reponsible for handling the pattern of watched lists.

This is a useful solution for working with entities that interact with each other.
For example, when creating a lesson with attachments, we might face challenges in
identifying what needs to be inserted, updated, or deleted during an update.

With this pattern and the accompanying code, it's easier to trach changes in our
entities, making database operations simpler to execute.